requirements.txt
google-genai
groq
matplotlib
openai
pydantic
pydantic[email]
python-dotenv
rich
streamlit
chromadb
sentence-transformers
rank-bm25
tiktoken
pypdf
numpy
langchain_openai
langchain_core
langgraph
supabase
psycopg2-binary
sqlparse
fpdf2 